+++ back/test/email.delegate.test.js
@@ -1,0 +1,31 @@
+const assert = require('assert');
+const Delegate = require('../routes/email.delegate');
+
+describe('email.delegate testing resolve/reject behavior ', () => {
+  var mockService = {
+    sendEmail: function(flag) {
+      if (flag) {
+        return Promise.resolve();
+      } else {
+        return Promise.reject('failed');
+      }
+      
+    }    
+  };
+
+  var delegate = new Delegate({emailService: mockService});
+
+  it('should return success', () => {
+    return delegate.sendEmail(1)
+      .then(response => {
+        assert.equal(response.success, true)
+      })
+  });
+
+  it('should return false', () => {
+    return delegate.sendEmail(0)
+      .then(response => {
+        assert.equal(response.success, false)
+      })
+  });
+})