+++ front/src/app/shared/material.module.ts
@@ -1,4 +1,5 @@
 import { NgModule } from '@angular/core';
+import {BidiModule} from '@angular/cdk/bidi';
 import {
   MatButtonModule,
   MatCardModule,
@@ -7,12 +8,14 @@ import {
   MatInputModule,
   MatListModule,
   MatSidenavModule,
+  MatSnackBarModule,
   MatToolbarModule
  } from '@angular/material';
 import { FlexLayoutModule } from '@angular/flex-layout';
 
 @NgModule({
   exports: [
+    BidiModule,
     MatButtonModule,
     MatCardModule,
     MatFormFieldModule,
@@ -20,6 +23,7 @@ import { FlexLayoutModule } from '@angular/flex-layout';
     MatInputModule,
     MatListModule,
     MatSidenavModule,
+    MatSnackBarModule,
     MatToolbarModule,
     FlexLayoutModule
   ]